Fifteen HID gamepad slots, which is all the selector nibble has
The HID select register at $FFB0 carries the device class in its low
nibble and the pad index in its high one, and index 0 is the firmware's
merged "any pad" view, so fifteen pads are addressable, not sixteen.

RIA816_PAD_SLOTS only follows the firmware's PAD_MAX_PLAYERS, which has
to be raised to match: ria816.c includes north/hid/pad.c outright, so a
real gamepad past the fourth is refused at mount until it is. Injected
pads are served from this side and work either way.

pad_inject_mask was a uint8_t indexed by `1u << (slot - 1)`, so slots
nine and up would have fallen off the end of it.

+++ src/chips/ria816.c
@@ -275,7 +275,7 @@ static void pad_synth_report(pad_connection_t* conn, void* data, uint16_t event_
 #define PAD_CONNECTED_BIT 0x80
 
 static uint8_t pad_inject_regs[RIA816_PAD_SLOTS][RIA816_PAD_REGS];
-static uint8_t pad_inject_mask;  // bit n set: slot n is scripted
+static uint16_t pad_inject_mask; // bit n set: slot n is scripted
 
 void ria816_pad_inject(uint8_t pad, const uint8_t report[RIA816_PAD_REGS]) {
     if (pad < 1 || pad > RIA816_PAD_SLOTS) return;
